Added a flag to include an additional listen metrics url
There is an issue in establishing communication between etcd endpoints and monitoring prometheus to do queries to the /metrics endpoint.

So, this change adds/exposes an additional url to be used as url for metrics queries.

This is related to this issue:
rackerlabs/kaas#1886

and to this PR:
rackerlabs/kaas#1985